# packages/rest-api/openapi.yaml
openapi: 3.0.3
info:
  title: EU Regulations Compliance API
  version: 0.4.1
  description: |
    Searchable access to 37 EU cybersecurity and data protection regulations.
    Supports full-text search, article retrieval, control framework mappings (ISO 27001, NIST CSF),
    and sector applicability rules.

    **Key Features:**
    - 2,278 articles across 37 regulations (GDPR, NIS2, DORA, AI Act, CRA, etc.)
    - 3,508 recitals with contextual guidance
    - 1,121 official definitions
    - 686 control mappings (ISO 27001:2022, NIST CSF 2.0)
    - 305 sector applicability rules

  contact:
    name: Ansvar Systems
    email: hello@ansvar.ai
    url: https://ansvar.ai
  license:
    name: MIT
    url: https://github.com/Ansvar-Systems/EU_compliance_MCP/blob/main/LICENSE
servers:
  - url: https://eu-regulations-api.ansvar.ai
    description: Production server
  - url: https://staging-eu-regulations-api.ansvar.ai
    description: Staging server
  - url: http://localhost:3000
    description: Local development
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/articles/{regulation}/{articleNumber}:
    get:
      summary: Get specific article
      description: Retrieve a specific article by regulation ID and article number
      operationId: getArticle
      tags:
        - articles
      parameters:
        - name: regulation
          in: path
          required: true
          description: Regulation ID (e.g., GDPR, NIS2, DORA)
          schema:
            type: string
            example: GDPR
        - name: articleNumber
          in: path
          required: true
          description: Article number (can include letters, e.g., "17", "5a")
          schema:
            type: string
            example: "17"
      responses:
        '200':
          description: Article found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/articles/{regulation}:
    get:
      summary: List all articles for a regulation
      description: Get all articles for a specific regulation with basic metadata
      operationId: listArticles
      tags:
        - articles
      parameters:
        - name: regulation
          in: path
          required: true
          description: Regulation ID
          schema:
            type: string
            example: NIS2
      responses:
        '200':
          description: List of articles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Microsoft Entra ID (Azure AD) JWT token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
  schemas:
...
    ArticleResponse:
      type: object
      properties:
        regulation:
          type: string
        article:
          type: string
        title:
          type: string
        text:
          type: string
        chapter:
          type: string
        recitals:
          type: array
          items:
            type: integer
          description: Related recital numbers
        cross_references:
          type: string
          nullable: true
    ArticleListResponse:
      type: object
      properties:
        regulation:
          type: string
        count:
          type: integer
        articles:
          type: array
          items:
            type: object
            properties:
              number:
                type: string
              title:
                type: string
              chapter:
                type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Invalid or expired token
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not Found
            message: Article 999 not found in GDPR